project-vib-sim/requirements.txt
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
streamlit>=1.28.0
plotly>=5.17.0
pyyaml>=6.0
networkx>=3.1
pytest>=7.4.0



